// frontend/src/features/auth/utils/enums.ts
export enum INPUT_VALUES_ENUM {
  EMAIL = "EMAIL",
  PASSWORD = "PASSWORD",
  REPEAT_PASSWORD = "REPEAT_PASSWORD",
  PRIVACY_POLICY = "PRIVACY_POLICY",
}

export enum AUTH_MODE_ENUM {
  LOGIN = "LOGIN",
  REGISTER = "REGISTER",
  FORGOT_PASSWORD = "FORGOT_PASSWORD",
  CHANGE_FORGOTTEN_PASSWORD = "CHANGE_FORGOTTEN_PASSWORD",
}

export enum VALIDATION_STATUS_ENUM {
  ERROR = "ERROR",
  SUCCESS = "SUCCESS",
}

export enum AUTH_RESPONSE_ENUM {
  ERROR = "ERROR",
  SUCCESS = "SUCCESS",
  EMAIL_VERIFICATION_REQUIRED = "EMAIL_VERIFICATION_REQUIRED",
}